Moving promo apply to ticket page and setting a loader
